+++ knet/src/main/kotlin/com/jasonernst/knet/ip/v4/options/Ipv4OptionLooseSourceAndRecordRoute.kt
@@ -1,73 +1,55 @@
-package com.jasonernst.knet.ip.options
+package com.jasonernst.knet.ip.v4.options
 
 import com.jasonernst.knet.PacketTooShortException
-import com.jasonernst.knet.ip.options.Ipv4OptionLooseSourceAndRecordRoute.Companion.MIN_OPTION_SIZE
 import org.slf4j.LoggerFactory
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
 
 /**
- * From the RFC 791:
+ * From RFC 791:
  *
- * The record route option provides a means to record the route of
- * an internet datagram.
+ * The loose source and record route (LSRR) option provides a means
+ * for the source of an internet datagram to supply routing
+ * information to be used by the gateways in forwarding the
+ * datagram to the destination, and to record the route
+ * information.
  *
  * The option begins with the option type code.  The second octet
  * is the option length which includes the option type code and the
  * length octet, the pointer octet, and length-3 octets of route
  * data.  The third octet is the pointer into the route data
- * indicating the octet which begins the next area to store a route
- * address.  The pointer is relative to this option, and the
+ * indicating the octet which begins the next source address to be
+ * processed.  The pointer is relative to this option, and the
  * smallest legal value for the pointer is 4.
  *
- * A recorded route is composed of a series of internet addresses.
+ * A route data is composed of a series of internet addresses.
  * Each internet address is 32 bits or 4 octets.  If the pointer is
- * greater than the length, the recorded route data area is full.
- * The originating host must compose this option with a large
- * enough route data area to hold all the address expected.  The
- * size of the option does not change due to adding addresses.  The
- * intitial contents of the route data area must be zero.
- *
- * When an internet module routes a datagram it checks to see if
- * the record route option is present.  If it is, it inserts its
- * own internet address as known in the environment into which this
- * datagram is being forwarded into the recorded route begining at
- * the octet indicated by the pointer, and increments the pointer
- * by four.
-
- * If the route data area is already full (the pointer exceeds the
- * length) the datagram is forwarded without inserting the address
- * into the recorded route.  If there is some room but not enough
- * room for a full address to be inserted, the original datagram is
- * considered to be in error and is discarded.  In either case an
- * ICMP parameter problem message may be sent to the source
- * host [3].
-
- * Not copied on fragmentation, goes in first fragment only.
- * Appears at most once in a datagram.
+ * greater than the length, the source route is empty (and the
+ * recorded route full) and the routing is to be based on the
+ * destination address field.
  */
-data class Ipv4OptionRecordRoute(
-    override val isCopied: Boolean = false,
+data class Ipv4OptionLooseSourceAndRecordRoute(
+    override val isCopied: Boolean = true,
     override val optionClass: Ipv4OptionClassType = Ipv4OptionClassType.Control,
-    override val type: Ipv4OptionType = Ipv4OptionType.RecordRoute,
+    override val type: Ipv4OptionType = Ipv4OptionType.LooseSourceRouting,
     val pointer: UByte,
     val routeData: ByteArray = ByteArray(0),
-) : Ipv4Option(isCopied = isCopied, optionClass = optionClass, type = type, size = (MIN_OPTION_SIZE + routeData.size.toUByte()).toUByte()) {
+) : Ipv4Option(isCopied = isCopied, optionClass = optionClass, type = type, size = (routeData.size.toUByte() + MIN_OPTION_SIZE).toUByte()) {
     companion object {
         val MIN_OPTION_SIZE: UByte = 3u
-        private val logger = LoggerFactory.getLogger(Ipv4OptionRecordRoute::class.java)
+        private val logger = LoggerFactory.getLogger(Ipv4OptionLooseSourceAndRecordRoute::class.java)
 
         fun fromStream(
             stream: ByteBuffer,
             isCopied: Boolean,
             optionClass: Ipv4OptionClassType,
             size: UByte,
-        ): Ipv4OptionRecordRoute {
+        ): Ipv4OptionLooseSourceAndRecordRoute {
             logger.debug("SIZE: $size, remaining: ${stream.remaining()}")
             if (stream.remaining() < (size - 2u).toInt()) {
                 throw PacketTooShortException(
                     "Stream must have at least ${size - 2u} " +
-                        "remaining bytes remaining to parse Ipv4OptionRecordRoute, we only have " +
+                        "remaining bytes remaining to parse Ipv4OptionLooseSourceAndRecordRoute, we only have " +
                         "${stream.remaining()} bytes",
                 )
             }
@@ -75,7 +57,7 @@ data class Ipv4OptionRecordRoute(
             val dataLength = size.toInt() - MIN_OPTION_SIZE.toInt()
             val routingData = ByteArray(dataLength)
             stream.get(routingData)
-            return Ipv4OptionRecordRoute(
+            return Ipv4OptionLooseSourceAndRecordRoute(
                 isCopied = isCopied,
                 optionClass = optionClass,
                 pointer = pointer,
@@ -99,7 +81,7 @@ data class Ipv4OptionRecordRoute(
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
 
-        other as Ipv4OptionRecordRoute
+        other as Ipv4OptionLooseSourceAndRecordRoute
 
         if (isCopied != other.isCopied) return false
         if (optionClass != other.optionClass) return false
