+++ knet/src/main/kotlin/com/jasonernst/knet/ip/v4/options/Ipv4OptionSecurityType.kt
@@ -1,4 +1,4 @@
-package com.jasonernst.knet.ip.options
+package com.jasonernst.knet.ip.v4.options
 
 enum class Ipv4OptionSecurityType(
     val kind: UShort,
