+++ knet/src/main/kotlin/com/jasonernst/knet/ip/v4/options/Ipv4OptionStrictSourceAndRecordRoute.kt
@@ -1,4 +1,4 @@
-package com.jasonernst.knet.ip.options
+package com.jasonernst.knet.ip.v4.options
 
 import com.jasonernst.knet.PacketTooShortException
 import org.slf4j.LoggerFactory
