+++ knet/src/main/kotlin/com/jasonernst/knet/ip/v6/Ipv6Header.kt
@@ -1,7 +1,10 @@
-package com.jasonernst.knet.ip
+package com.jasonernst.knet.ip.v6
 
 import com.jasonernst.knet.PacketTooShortException
+import com.jasonernst.knet.ip.IpHeader
 import com.jasonernst.knet.ip.IpHeader.Companion.IP6_VERSION
+import com.jasonernst.knet.ip.IpType
+import com.jasonernst.knet.ip.v6.extenions.Ipv6ExtensionHeader
 import org.slf4j.LoggerFactory
 import java.net.Inet6Address
 import java.net.InetAddress
@@ -72,7 +75,9 @@ data class Ipv6Header(
             val destinationBuffer = ByteArray(16)
             stream[destinationBuffer]
             val destinationAddress = Inet6Address.getByAddress(destinationBuffer) as Inet6Address
-            val extensionHeaders = Ipv6ExtensionHeader.fromStream(stream, IpType.fromValue(protocol))
+            val extensionHeaders = Ipv6ExtensionHeader.fromStream(stream,
+                IpType.fromValue(protocol)
+            )
 
             return Ipv6Header(
                 ipVersion,
