+++ knet/src/test/kotlin/com/jasonernst/knet/PacketTests.kt
@@ -2,8 +2,8 @@ package com.jasonernst.knet
 
 import com.jasonernst.knet.ip.IpHeader
 import com.jasonernst.knet.ip.IpType
-import com.jasonernst.knet.ip.Ipv4Header
-import com.jasonernst.knet.ip.Ipv6Header
+import com.jasonernst.knet.ip.v4.Ipv4Header
+import com.jasonernst.knet.ip.v6.Ipv6Header
 import com.jasonernst.knet.transport.tcp.TcpHeader
 import com.jasonernst.knet.transport.udp.UdpHeader
 import org.junit.jupiter.api.Assertions.assertEquals
