+++ knet/src/test/kotlin/com/jasonernst/knet/transport/tcp/options/TcpOptionTests.kt
@@ -3,7 +3,7 @@ package com.jasonernst.knet.tcp.options
 import com.jasonernst.knet.PacketTooShortException
 import com.jasonernst.knet.ip.IpHeader
 import com.jasonernst.knet.ip.IpType
-import com.jasonernst.knet.ip.options.Ipv4OptionNoOperation
+import com.jasonernst.knet.ip.v4.options.Ipv4OptionNoOperation
 import com.jasonernst.knet.nextheader.NextHeader
 import com.jasonernst.knet.transport.tcp.TcpHeader
 import com.jasonernst.knet.transport.tcp.options.TcpOption
